refactor: move to ape.utils.misc
