+++ src/ape/utils/__init__.py
@@ -30,6 +30,7 @@ from ape.utils.misc import (
     gas_estimation_error_message,
     get_package_version,
     load_config,
+    nonreentrant,
     raises_not_implemented,
     run_until_complete,
     singledispatchmethod,
@@ -84,6 +85,7 @@ __all__ = [
     "load_config",
     "LogInputABICollection",
     "ManagerAccessMixin",
+    "nonreentrant",
     "parse_gas_table",
     "raises_not_implemented",
     "returns_array",
